Corrected Division
I was dividing Interval Time by Interval Length, forgetting that Interval Time is greater than 1440 minutes ( a single day), after the 1st repetition. Interval Length needed to be divided by 1440 minutes every time, so the quotient would be the number of days to add to the current date of the study session. So I corrected it to do that.